Reshuffled the application files
