+++ src/frontend/js/views/App/CommandBar.jsx
