+++ src/frontend/js/views/App/JobProgress.jsx
