+++ src/frontend/js/views/App/LeftCommandBar.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import CommandBar from 'views/CommandBar.jsx'
+import CommandBar from 'views/App/CommandBar.jsx'
 
 class LeftCommandBar extends React.Component {
     constructor(props) {
