+++ src/frontend/js/views/App/Pane/PaneFile.jsx
