+++ src/frontend/js/views/App/Pane/PaneHeader.jsx
