+++ src/frontend/js/views/App/StatusBar.jsx
