+++ src/frontend/js/views/Login.jsx
@@ -1,0 +1,33 @@
+import React from 'react';
+
+class Login extends React.Component {
+    constructor(props) {
+        super(props);
+    }
+
+    render() {
+        return (
+            <div>
+                Login
+            </div>
+        );
+    }
+
+    componentDidMount() {
+
+    }
+}
+
+Login.defaultProps = {
+
+}
+
+import {connect} from 'react-redux';
+
+const mapStateToProps = state => ({
+});
+
+const mapDispatchToProps = dispatch => ({
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(Login);
